♻️ refactor: context 위치 변경
