+++ src/App.tsx
@@ -1,9 +1,9 @@
 import { BrowserRouter, Route, Routes } from 'react-router';
 import CartPage from './pages/cart/CartPage';
 import OrderConfirmPage from './pages/orderConfirm/OrderConfirmPage';
-import { CartProvider } from './global/contexts/CartContext';
-import { CartSelectionProvider } from './global/contexts/CartSelectionContext';
+import { CartSelectionProvider } from './components/features/cart/contexts/CartSelectionContext';
 import PaymentPage from './pages/payment/PaymentPage';
+import { CartProvider } from './components/features/cart/contexts/CartContext';
 
 function App() {
   return (
